Renames SeekApp to StyleGuideProvider (with b/c alias)
